Fixed compilation errors and warnings

+++ Packages/Parameterization/test/Parameterization/extensive_parameterization_test.C
@@ -240,8 +240,8 @@ int main(int argc,char * argv[])
         stream >> mesh;
 
         // print mesh info
-        fprintf(stderr, "(%d facets, ",mesh.size_of_facets());
-        fprintf(stderr, "%d vertices)\n",mesh.size_of_vertices());
+        fprintf(stderr, "(%d facets, ", (int)mesh.size_of_facets());
+        fprintf(stderr, "%d vertices)\n", (int)mesh.size_of_vertices());
 
         //***************************************
         // Create mesh adaptor
